moved document calls to DocumentServices, remove console.log()s
