+++ app/src/services/ProjectServices.ts
@@ -5,11 +5,9 @@ export const projectServices = {
     getProjectNames,
     getProjectUsers,
     setProjectUsers,
-    setProjectUsers,
     getProjectTags,
     setProjectTags,
-    getDocument,
-    getLabels,
+
 }
 
 async function getProjectNames(firebase: any) {
@@ -147,37 +145,6 @@ async function getProjectUsers(project: string, firebase: any) {
             return data.users
         })
  }
-
- function getDocument(project_name: any, document_id: any, firebase: any) {
-
-
-   const requestOptions = {
-       method: 'GET',
-       headers: { 'Content-Type': 'application/json' },
-   };
-   return fetch(process.env.REACT_APP_API_URL + '/projects/' + project_name + '/documents/' + document_id + '?id_token=' + localStorage.getItem('user-token'), requestOptions) // TODO:config.apiUrl
-       .then(handleResponse)
-       .then(data => {
-
-           return data.document
-       })
-}
-
-function getLabels(project_name: any, firebase: any) {
-
-
-   const requestOptions = {
-       method: 'GET',
-       headers: { 'Content-Type': 'application/json' },
-   };
-   return fetch(process.env.REACT_APP_API_URL + '/projects/' + project_name + '/labels/all' + '?id_token=' + localStorage.getItem('user-token'), requestOptions) // TODO:config.apiUrl
-       .then(handleResponse)
-       .then(data => {
-
-           return data.labels
-       })
-}
-
 
 function handleResponse(response: { text: () => Promise<any>; ok: any; status: number; statusText: any; }) {
    return response.text().then((text: string) => {
